3rd featurebrunch1 commit

+++ src/Wheels.java
@@ -1,3 +1,3 @@
 public class Wheels {
-    private int diametr;
+    private double diametr;
 }
